+++ src/run_agent_interactive.py
@@ -11,7 +11,7 @@ from run_agent import run_experiment
 from utils.config import Config
 from utils.dataset_utils import get_all_prepared_datasets_info, get_task_type_from_prepared_dataset
 from utils.datasets_interactive_utils import interactive_dataset_selection, print_datasets_table
-from utils.env_utils import are_wandb_vars_available
+from run_logging.env_utils import are_wandb_vars_available
 from utils.metrics import get_classification_metrics_names, get_regression_metrics_names, resolve_val_metric
 from utils.metrics_interactive_utils import display_metrics_table
 from utils.providers.provider import Provider, get_provider_and_api_key
